requirements.txt
openai>=1.40.0
requests>=2.32.0
python-dotenv>=1.0.1
PyYAML>=6.0.1
Pillow>=10.3.0
